remove SecondaryNavbar component; consolidate into one
